+++ zray/DoctrineOrm.php
@@ -22,9 +22,6 @@ use Doctrine\ORM\Mapping\ClassMetadata;
  */
 class DoctrineOrm
 {
-    const INDEX_ENTITIES_UNIQUE = 'unique_entities';
-    const INDEX_ENTITIES_REFERENCE = 'referenced_entities';
-
     /**
      * Entities
      *
@@ -178,36 +175,6 @@ class DoctrineOrm
     }
 
     /**
-     * Collects queries from \Doctrine\ORM\Persisters\Entity\BasicEntityPersister
-     *
-     * @param array $context
-     * @param array $storage
-     */
-    public function persister($context, &$storage)
-    {
-        // dont break z-ray
-        if (empty($context['functionArgs'][0])
-            || !$this->lastQuery
-            || empty($context['returnValue'])
-        ) {
-            $this->lastQuery = '';
-            return;
-        }
-
-        $params = [];
-
-        foreach ($context['returnValue'] as &$values) {
-            $params += $values;
-        }
-
-        $this->queries[$this->lastQuery]['query'] = $this->formatQuery(
-            $this->queries[$this->lastQuery]['query'],
-            $params
-        );
-        $this->lastQuery = '';
-    }
-
-    /**
      * Collects all queries from Doctrine\DBAL\Connection
      *
      * @param array $context
@@ -269,6 +236,9 @@ class DoctrineOrm
             return;
         }
 
+        $query = '';
+        $queryId = '';
+
         switch ($context['functionName']) {
             case 'Doctrine\DBAL\Connection::beginTransaction':
                 $query = 'Begin Transaction';
@@ -285,15 +255,45 @@ class DoctrineOrm
                 $this->queryNumber++;
                 $queryId = $query . $this->queryNumber;
                 break;
-            case 'Doctrine\DBAL\Statement::__construct':
+            case 'Doctrine\DBAL\Statement::execute':
+                $reflection = new \ReflectionObject($context['this']);
+
+                if (!$reflection->hasProperty('params')
+                    || !$reflection->hasProperty('types')
+                    || !$reflection->hasProperty('sql')
+                ) {
+                    break;
+                }
+
+                if (empty($context['functionArgs'][0])) {
+                    $params = $reflection->getProperty('params');
+                    $params->setAccessible(true);
+                    $params = $params->getValue($context['this']);
+                } else {
+                    $params = $context['functionArgs'][0];
+                }
+
+                $types = $reflection->getProperty('types');
+                $types->setAccessible(true);
+                $types = $types->getValue($context['this']);
+
+                $query = $reflection->getProperty('sql');
+                $query->setAccessible(true);
+                $query = $query->getValue($context['this']);
+
+                # doctrine does it right
+                list($queryWithParams, $params, $types) = \Doctrine\DBAL\SQLParserUtils::expandListParameters(
+                    $query,
+                    $params,
+                    $types
+                );
+
+                $query = $this->formatQuery($queryWithParams, $params);
                 $this->queryNumber++;
-                $query = isset($context['locals']['sql']) ? $context['locals']['sql'] : 'NO QUERY';
                 $queryId = $query . $this->queryNumber;
-                $this->lastQuery = $this->getQueryId($queryId);
                 break;
             default:
-                $query = '';
-                $queryId = '';
+                // nothing to do
                 break;
         }
 
@@ -402,10 +402,18 @@ class DoctrineOrm
         $this->events[$eventName]['listeners'][$listener] = true;
     }
 
+    /**
+     * Collects listener from Doctrine EventManager
+     *
+     * @param array $context
+     * @param array $storage
+     */
     public function eventManagerAddListener($context, &$storage)
     {
         // dont break z-ray
-        if (empty($context['functionArgs'][0])) {
+        if (empty($context['functionArgs'][0])
+            || empty($context['functionArgs'][1])
+        ) {
             return;
         }
 
@@ -428,7 +436,6 @@ class DoctrineOrm
                 $this->events[$event]['listeners'][$listener] = true;
             }
         }
-
     }
 
     /**